+++ script/check_i18n_catalogs.py
@@ -1,4 +1,6 @@
+#!/usr/bin/env python3
 # SPDX-License-Identifier: MIT
+# Copyright (c) 2021-2025
 """
 Pre-commit hook to validate i18n catalogs and usage.