+++ script/sort_class_members.py
@@ -1,4 +1,6 @@
 #!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+# Copyright (c) 2021-2025
 """
 Auto-fix pre-commit helper to reorder Python class members.